Added support for chownat() and fstatat()
Included fstatat() support which in fact requires
__fxstatat() and __fxstatat64(). Just like __xstat() and its relatives,
they are the back-end for fstatat(), which in glibc is implemented as
an inline function. See the details in sys/stat.h.

While doing a strace of the chown utility I found out it calls chmod32()
so I added a small hook to support it.