work around jemalloc's config file lookup
If a binary is linked with jemalloc, installwatch will
invoke dlsym from libdl, which will call calloc
in jemalloc, which will initialize the jemalloc data
structures, which in turn tries to look at /etc/malloc.conf,
which is apparently no optional behavior.

This is a dirty workaround that catches this call and denies it.